pro/requirements.txt
# Core web framework
Flask==2.3.3
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-OAuthlib==0.9.6
Werkzeug==2.3.7
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
click==8.1.7
blinker==1.6.2

# Database and ORM
SQLAlchemy==2.0.20
alembic==1.12.0
pymysql==1.1.0
psycopg2-binary==2.9.7
mysqlclient==2.2.0
sqlite3-api==0.1.0

# Data processing and analysis
pandas==2.0.3
numpy==1.24.4
scipy==1.11.2
pyarrow==12.0.1
openpyxl==3.1.2
xlrd==2.0.1
xlwt==1.3.0
xlsxwriter==3.1.2
fastparquet==2023.7.0
tabulate==0.9.0
beautifulsoup4==4.12.2
lxml==4.9.3
html5lib==1.1
python-dateutil==2.8.2

# Machine learning
scikit-learn==1.3.0
tensorflow==2.13.0
keras==2.13.1
torch==2.0.1
torchvision==0.15.2
xgboost==1.7.6
lightgbm==4.0.0
catboost==1.2
statsmodels==0.14.0
joblib==1.3.2
imbalanced-learn==0.11.0
category_encoders==2.6.1
shap==0.42.1
lime==0.2.0.1
eli5==0.13.0
sktime==0.20.0
Prophet==1.1.4
pmdarima==2.0.3

# Quantum computing
pennylane==0.30.0
qiskit==0.44.0
cirq==1.2.0
pyquil==3.5.4

# Visualization
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.16.1
bokeh==3.2.2
pygal==3.0.0
altair==5.0.1
holoviews==1.17.1
graphviz==0.20.1
pydot==1.4.2
networkx==3.1

# PDF and report generation
reportlab==3.6.13
pdfkit==1.0.0
weasyprint==59.0
xhtml2pdf==0.2.11
fpdf==1.7.2
pypdf2==3.0.1
pdfminer.six==20221105

# Image processing
pillow==10.0.0
opencv-python==4.8.0.76
scikit-image==0.21.0
imageio==2.31.1
albumentations==1.3.1

# API and networking
requests==2.31.0
httpx==0.24.1
aiohttp==3.8.5
urllib3==2.0.4
websockets==11.0.3
socketio==0.2.1
flask-socketio==5.3.4
gunicorn==21.2.0
uvicorn==0.23.2
werkzeug==2.3.7

# Authentication and security
pyotp==2.9.0
passlib==1.7.4
bcrypt==4.0.1
pyjwt==2.8.0
cryptography==41.0.3
oauthlib==3.2.2
python-jose==3.3.0
authlib==1.2.1

# AI and NLP
google-generativeai==0.3.1
transformers==4.32.1
sentence-transformers==2.2.2
spacy==3.6.1
nltk==3.8.1
gensim==4.3.1
openai==0.28.0
langchain==0.0.285
haystack==1.20.1
bert-embedding==1.0.1

# Cloud services and storage
boto3==1.28.36
google-cloud-storage==2.10.0
azure-storage-blob==12.17.0
firebase-admin==6.2.0

# Development and testing
python-dotenv==1.0.0
pytest==7.4.0
pytest-flask==1.2.0
pytest-cov==4.1.0
coverage==7.3.1
mock==5.1.0
faker==19.3.1
black==23.7.0
isort==5.12.0
flake8==6.1.0
mypy==1.5.1
pytest-mock==3.11.1

# Utilities and helpers
pyyaml==6.0.1
tqdm==4.66.1
colorama==0.4.6
rich==13.5.2
loguru==0.7.0
humanize==4.8.0
bleach==6.0.0
watchdog==3.0.0
cachetools==5.3.1
apscheduler==3.10.4
celery==5.3.1
redis==4.6.0
pymongo==4.5.0
jupyter==1.0.0
ipython==8.14.0
papermill==2.4.0
flask-caching==2.0.2
flask-restful==0.3.10
flask-cors==4.0.0
Flask==2.3.3 
